First attempt at trying to implement temporary passwords
I have added my first attempt at implementing the temporary passwords.
FirstLog is the boolean I use to track the first log in to an account. I
haven't added it to the database yet but I think I will have to in order
to check it from the other files. I have also created methods and
classes for updating the advisor information in the database, but I
don't think they will work in their current state.